Issue #14758 move ConnectionFactoryDefinition to separate servlet so it doesn't impact other test server
