+++ dev/com.ibm.ws.jca_fat_derbyra/fat/src/com/ibm/ws/jca/fat/derbyra/DerbyResourceAdapterTest.java
@@ -38,6 +38,7 @@ public class DerbyResourceAdapterTest extends FATServletClient {
 
     private static final String derbyRAAppName = "derbyRAAppName";
     private static final String DerbyRAAnnoServlet = "fvtweb/DerbyRAAnnoServlet";
+    private static final String DerbyRACFDServlet = "fvtweb/DerbyRACFDServlet";
     private static final String DerbyRAServlet = "fvtweb/DerbyRAServlet";
 
     @Server("com.ibm.ws.jca.fat.derbyra")
@@ -48,6 +49,7 @@ public class DerbyResourceAdapterTest extends FATServletClient {
 
         WebArchive war = ShrinkWrap.create(WebArchive.class, WAR_NAME + ".war");
         war.addPackage("web");
+        war.addPackage("web.cfd");
         war.addPackage("web.mdb");
         war.addAsWebInfResource(new File("test-applications/fvtweb/resources/WEB-INF/ibm-web-bnd.xml"));
         war.addAsWebInfResource(new File("test-applications/fvtweb/resources/WEB-INF/web.xml"));
@@ -131,14 +133,14 @@ public class DerbyResourceAdapterTest extends FATServletClient {
 
     @Test
     public void testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseEnabled() throws Exception {
-        runTest(server, DerbyRAAnnoServlet, "testConnectionFactoryDefinitionLeakConnectionWithAutoCloseEnabled");
-        runTest(server, DerbyRAAnnoServlet, "testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseEnabledClosed");
+        runTest(server, DerbyRACFDServlet, "testConnectionFactoryDefinitionLeakConnectionWithAutoCloseEnabled");
+        runTest(server, DerbyRACFDServlet, "testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseEnabledClosed");
     }
 
     @Test
     public void testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseDisabled() throws Exception {
-        runTest(server, DerbyRAAnnoServlet, "testConnectionFactoryDefinitionLeakConnectionWithAutoCloseDisabled");
-        runTest(server, DerbyRAAnnoServlet, "testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseDisabledNotClosed");
+        runTest(server, DerbyRACFDServlet, "testConnectionFactoryDefinitionLeakConnectionWithAutoCloseDisabled");
+        runTest(server, DerbyRACFDServlet, "testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseDisabledNotClosed");
     }
 
     @Test
