+++ dev/com.ibm.ws.jca_fat_derbyra/test-applications/fvtweb/src/web/DerbyRAAnnoServlet.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2017,2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -33,8 +33,6 @@ import java.util.concurrent.atomic.AtomicLong;
 
 import javax.annotation.Resource;
 import javax.naming.InitialContext;
-import javax.resource.ConnectionFactoryDefinition;
-import javax.resource.ConnectionFactoryDefinitions;
 import javax.resource.spi.BootstrapContext;
 import javax.resource.spi.XATerminator;
 import javax.resource.spi.work.ExecutionContext;
@@ -53,18 +51,6 @@ import javax.transaction.xa.Xid;
 
 import componenttest.app.FATServlet;
 
-@ConnectionFactoryDefinitions({
-                                @ConnectionFactoryDefinition(
-                                                             name = "java:module/env/eis/ds6-auto-close-true",
-                                                             interfaceName = "javax.sql.DataSource",
-                                                             resourceAdapter = "DerbyRA",
-                                                             properties = "autoCloseConnections=true"),
-                                @ConnectionFactoryDefinition(
-                                                             name = "java:module/env/eis/ds7-auto-close-false",
-                                                             interfaceName = "javax.sql.DataSource",
-                                                             resourceAdapter = "DerbyRA",
-                                                             properties = "autoCloseConnections=false")
-})
 public class DerbyRAAnnoServlet extends FATServlet {
 
     private static final long serialVersionUID = 7440572626782340872L;
@@ -72,20 +58,11 @@ public class DerbyRAAnnoServlet extends FATServlet {
     @Resource(name = "java:global/env/eis/bootstrapContext", lookup = "eis/bootstrapContext")
     private BootstrapContext bootstrapContext;
 
-    // Intentionally left open by test after servlet method ends. Never do this in a real application.
-    private static Connection cachedConnection;
-
     @Resource(lookup = "eis/ds1")
     private CommonDataSource cds1;
 
     @Resource(name = "java:module/env/eis/ds1ref", lookup = "eis/ds1")
     private DataSource ds1;
-
-    @Resource(lookup = "java:module/env/eis/ds6-auto-close-true", shareable = false)
-    private DataSource ds6;
-
-    @Resource(lookup = "java:module/env/eis/ds7-auto-close-false", shareable = false)
-    private DataSource ds7;
 
     @Resource(name = "java:app/env/eis/map1ref", lookup = "eis/map1")
     private Map<String, String> map1;
@@ -199,55 +176,6 @@ public class DerbyRAAnnoServlet extends FATServlet {
             assertEquals("mdbtestRecovery: valueA --> valueB", oldValueFromDB);
         } finally {
             map1.clear();
-        }
-    }
-
-    /**
-     * First part of a test that leaves a connection handle open across the end of a servlet request
-     * and expects the container to automatically close it.
-     */
-    public void testConnectionFactoryDefinitionLeakConnectionWithAutoCloseEnabled() throws Throwable {
-        cachedConnection = ds6.getConnection();
-    }
-
-    /**
-     * First part of a test that leaves a connection handle open across the end of a servlet request
-     * and expects the container to leave it open.
-     */
-    public void testConnectionFactoryDefinitionLeakConnectionWithAutoCloseDisabled() throws Throwable {
-        cachedConnection = ds7.getConnection();
-    }
-
-    /**
-     * Second part of a test that leaves a connection handle open across the end of a servlet request
-     * and expects the container to automatically close it.
-     */
-    public void testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseEnabledClosed() throws Throwable {
-        Connection con = cachedConnection;
-        try {
-            cachedConnection = null;
-            assertTrue(con.isClosed());
-        } finally {
-            con.close();
-        }
-    }
-
-    /**
-     * Second part of a test that leaves a connection handle open across the end of a servlet request
-     * and expects the container to leave it open. The connection must still be usable.
-     */
-    public void testConnectionFactoryDefinitionLeakedConnectionWithAutoCloseDisabledNotClosed() throws Throwable {
-        Connection con = cachedConnection;
-        try {
-            ResultSet result = con.createStatement().executeQuery("VALUES(7)");
-            assertTrue(result.next());
-            assertEquals(7, result.getInt(1));
-            Statement st = result.getStatement();
-            result.close();
-            st.close();
-        } finally {
-            cachedConnection = null;
-            con.close();
         }
     }
